+++ cms/djangoapps/coursegraph/management/commands/tests/test_dump_to_neo4j.py
@@ -13,10 +13,10 @@ from xmodule.modulestore.tests.django_utils import SharedModuleStoreTestCase
 from xmodule.modulestore.tests.factories import CourseFactory, ItemFactory
 
 import openedx.core.djangoapps.content.block_structure.config as block_structure_config
-from openedx.core.djangoapps.content.block_structure.signals import update_block_structure_on_course_publish
-from openedx.core.djangoapps.coursegraph.management.commands.dump_to_neo4j import ModuleStoreSerializer
-from openedx.core.djangoapps.coursegraph.management.commands.tests.utils import MockGraph, MockNodeMatcher
-from openedx.core.djangoapps.coursegraph.tasks import (
+from cms.djangoapps.content.block_structure.signals import update_block_structure_on_course_publish
+from cms.djangoapps.coursegraph.management.commands.dump_to_neo4j import ModuleStoreSerializer
+from cms.djangoapps.coursegraph.management.commands.tests.utils import MockGraph, MockNodeMatcher
+from cms.djangoapps.coursegraph.tasks import (
     coerce_types,
     serialize_course,
     serialize_item,
@@ -115,8 +115,8 @@ class TestDumpToNeo4jCommand(TestDumpToNeo4jCommandBase):
     Tests for the dump to neo4j management command
     """
 
-    @mock.patch('openedx.core.djangoapps.coursegraph.tasks.NodeMatcher')
-    @mock.patch('openedx.core.djangoapps.coursegraph.tasks.Graph')
+    @mock.patch('cms.djangoapps.coursegraph.tasks.NodeMatcher')
+    @mock.patch('cms.djangoapps.coursegraph.tasks.Graph')
     @ddt.data(1, 2)
     def test_dump_specific_courses(self, number_of_courses, mock_graph_class, mock_matcher_class):
         """
@@ -140,8 +140,8 @@ class TestDumpToNeo4jCommand(TestDumpToNeo4jCommandBase):
             number_rollbacks=0
         )
 
-    @mock.patch('openedx.core.djangoapps.coursegraph.tasks.NodeMatcher')
-    @mock.patch('openedx.core.djangoapps.coursegraph.tasks.Graph')
+    @mock.patch('cms.djangoapps.coursegraph.tasks.NodeMatcher')
+    @mock.patch('cms.djangoapps.coursegraph.tasks.Graph')
     def test_dump_skip_course(self, mock_graph_class, mock_matcher_class):
         """
         Test that you can skip courses.
@@ -166,8 +166,8 @@ class TestDumpToNeo4jCommand(TestDumpToNeo4jCommandBase):
             number_rollbacks=0,
         )
 
-    @mock.patch('openedx.core.djangoapps.coursegraph.tasks.NodeMatcher')
-    @mock.patch('openedx.core.djangoapps.coursegraph.tasks.Graph')
+    @mock.patch('cms.djangoapps.coursegraph.tasks.NodeMatcher')
+    @mock.patch('cms.djangoapps.coursegraph.tasks.Graph')
     def test_dump_skip_beats_specifying(self, mock_graph_class, mock_matcher_class):
         """
         Test that if you skip and specify the same course, you'll skip it.
@@ -193,8 +193,8 @@ class TestDumpToNeo4jCommand(TestDumpToNeo4jCommandBase):
             number_rollbacks=0,
         )
 
-    @mock.patch('openedx.core.djangoapps.coursegraph.tasks.NodeMatcher')
-    @mock.patch('openedx.core.djangoapps.coursegraph.tasks.Graph')
+    @mock.patch('cms.djangoapps.coursegraph.tasks.NodeMatcher')
+    @mock.patch('cms.djangoapps.coursegraph.tasks.Graph')
     def test_dump_all_courses(self, mock_graph_class, mock_matcher_class):
         """
         Test if you don't specify which courses to dump, then you'll dump
@@ -395,8 +395,8 @@ class TestModuleStoreSerializer(TestDumpToNeo4jCommandBase):
         coerced_value = coerce_types(original_value)
         assert coerced_value == coerced_expected
 
-    @mock.patch('openedx.core.djangoapps.coursegraph.tasks.NodeMatcher')
-    @mock.patch('openedx.core.djangoapps.coursegraph.tasks.authenticate_and_create_graph')
+    @mock.patch('cms.djangoapps.coursegraph.tasks.NodeMatcher')
+    @mock.patch('cms.djangoapps.coursegraph.tasks.authenticate_and_create_graph')
     def test_dump_to_neo4j(self, mock_graph_constructor, mock_matcher_class):
         """
         Tests the dump_to_neo4j method works against a mock
@@ -423,8 +423,8 @@ class TestModuleStoreSerializer(TestDumpToNeo4jCommandBase):
         assert len(mock_graph.nodes) == 11
         self.assertCountEqual(submitted, self.course_strings)
 
-    @mock.patch('openedx.core.djangoapps.coursegraph.tasks.NodeMatcher')
-    @mock.patch('openedx.core.djangoapps.coursegraph.tasks.authenticate_and_create_graph')
+    @mock.patch('cms.djangoapps.coursegraph.tasks.NodeMatcher')
+    @mock.patch('cms.djangoapps.coursegraph.tasks.authenticate_and_create_graph')
     def test_dump_to_neo4j_rollback(self, mock_graph_constructor, mock_matcher_class):
         """
         Tests that the the dump_to_neo4j method handles the case where there's
@@ -447,8 +447,8 @@ class TestModuleStoreSerializer(TestDumpToNeo4jCommandBase):
 
         self.assertCountEqual(submitted, self.course_strings)
 
-    @mock.patch('openedx.core.djangoapps.coursegraph.tasks.NodeMatcher')
-    @mock.patch('openedx.core.djangoapps.coursegraph.tasks.authenticate_and_create_graph')
+    @mock.patch('cms.djangoapps.coursegraph.tasks.NodeMatcher')
+    @mock.patch('cms.djangoapps.coursegraph.tasks.authenticate_and_create_graph')
     @ddt.data((True, 2), (False, 0))
     @ddt.unpack
     def test_dump_to_neo4j_cache(
@@ -480,8 +480,8 @@ class TestModuleStoreSerializer(TestDumpToNeo4jCommandBase):
         )
         assert len(submitted) == expected_number_courses
 
-    @mock.patch('openedx.core.djangoapps.coursegraph.tasks.NodeMatcher')
-    @mock.patch('openedx.core.djangoapps.coursegraph.tasks.authenticate_and_create_graph')
+    @mock.patch('cms.djangoapps.coursegraph.tasks.NodeMatcher')
+    @mock.patch('cms.djangoapps.coursegraph.tasks.authenticate_and_create_graph')
     def test_dump_to_neo4j_published(self, mock_graph_constructor, mock_matcher_class):
         """
         Tests that we only dump those courses that have been published after
@@ -506,8 +506,8 @@ class TestModuleStoreSerializer(TestDumpToNeo4jCommandBase):
         assert len(submitted) == 1
         assert submitted[0] == str(self.course.id)
 
-    @mock.patch('openedx.core.djangoapps.coursegraph.tasks.get_course_last_published')
-    @mock.patch('openedx.core.djangoapps.coursegraph.tasks.get_command_last_run')
+    @mock.patch('cms.djangoapps.coursegraph.tasks.get_course_last_published')
+    @mock.patch('cms.djangoapps.coursegraph.tasks.get_command_last_run')
     @ddt.data(
         (
             str(datetime(2016, 3, 30)), str(datetime(2016, 3, 31)),
